Add getArticles function
getArticles gets the HTML for articles containing recipes and stores it in a database, along with the discovered recipe names and, if they differ, the expected names, and along with certain article/recipe attributes.  getArticles uses the local NYT Food and Drink website to get articles starting from 3/29/2006 and continuing into the past.